requirements.txt
Django==2.0.2
django-filter==1.1.0
djangorestframework==3.7.7
freeze==1.0.10
Markdown==2.6.11
Pillow==5.0.0
pkg-resources==0.0.0
pytz==2017.3
six==1.11.0
